Added update dialog box

+++ src/components/ConfirmBox/ConfirmBox.jsx
@@ -1,24 +1,46 @@
 import { NoteContent } from "../../NotesContent";
 import "./ConfirmBox.scss";
 
-const ConfirmBox = ({ setIsDeleteBtnClicked, noteKey }) => {
+const ConfirmBox = ({
+  IsDeleteBtnClicked,
+  setIsDeleteBtnClicked,
+  noteKey,
+  isUpdatingNote,
+  setGottenDate,
+  setGottenDetails,
+  setGottenTitle,
+  noteTitle,
+  noteDetails,
+  noteDate,
+}) => {
   const deleteNote = () => {
     let filteredNote = NoteContent.filter((note) => {
       return note.key === noteKey;
     });
 
     NoteContent.splice(filteredNote - 1, 1);
-    // window.location.reload();
+    window.location.reload();
+  };
+
+  const updateNote = () => {
+    setGottenTitle(noteTitle);
+    setGottenDetails(noteDetails);
+    setGottenDate(noteDate);
   };
 
   return (
     <div className="confirmBox">
-      <p className="confirmBox__text">Delete Active Note?</p>
+      {isUpdatingNote ? (
+        <p className="confirmBox__text">Update Note?</p>
+      ) : IsDeleteBtnClicked ? (
+        <p className="confirmBox__text">Delete Note?</p>
+      ) : null}
       <div className="confirmBox__btn">
         <button
           className="confirmBox__btn--1"
           onClick={() => {
-            deleteNote();
+            IsDeleteBtnClicked && deleteNote();
+            isUpdatingNote && updateNote();
           }}
         >
           Confirm
